Dashboard: move chunk-load recovery out of the logger folder
The runtime recovery logic reads and resets window.location, so it does
not belong under app/logger/. Move it up to app/chunk-load-recovery.ts.
